[220810]service dashboard

+++ gmc_api_gateway/app/model/dashboard.go
@@ -10,20 +10,19 @@ type TOTAL_DASHBOARD struct {
 	PodCpuTop5     interface{} `json:"podCpuTop5"`
 	ClusterMemTop5 interface{} `json:"clusterMemTop5"`
 	PodMemTop5     interface{} `json:"podMemTop5"`
-	CredentialCnt  int     	   `json:"credentialCnt"`
+	CredentialCnt  int         `json:"credentialCnt"`
 	EdgeCloud      interface{} `json:"edgeInfo"`
 }
 
 type CoreCloud struct {
-	
 }
 
 type EdgeCloud struct {
 }
 
 type SERVICE_DASHBOARD struct {
-	WorkspaceCnt   int64       `json:"workspaceCnt"`
-	ProjectCnt     int64       `json:"projectCnt"`
+	WorkspaceCnt   int         `json:"workspaceCnt"`
+	ProjectCnt     int         `json:"projectCnt"`
 	Resource       interface{} `json:"resource"`
 	PodCpuTop5     interface{} `json:"podCpuTop5"`
 	PodMemTop5     interface{} `json:"podMemTop5"`
@@ -31,20 +30,19 @@ type SERVICE_DASHBOARD struct {
 	ProjectMemTop5 interface{} `json:"projectMemTop5"`
 }
 
-type CLOUD_DASHBOARD struct { 
-	ClusterInfo      interface{} `json:"ClusterInfo"`
+type CLOUD_DASHBOARD struct {
+	ClusterInfo interface{} `json:"ClusterInfo"`
 
 	MemoryUsage interface{} `json:"memoryUsage"`
-	MemoryUtil interface{} `json:"memoryUtil"`
+	MemoryUtil  interface{} `json:"memoryUtil"`
 	MemoryTotal interface{} `json:"memoryTotal"`
-	CpuUsage interface{} `json:"cpuUsage"`
-	CpuUtil interface{} `json:"cpuUtil"`
-	CpuTotal interface{} `json:"cpuTotal"`
-	DiskUsage interface{} `json:"diskUsage"`
-	DiskUtil interface{} `json:"diskUtil"`
-	DiskTotal interface{} `json:"diskTotal"`
+	CpuUsage    interface{} `json:"cpuUsage"`
+	CpuUtil     interface{} `json:"cpuUtil"`
+	CpuTotal    interface{} `json:"cpuTotal"`
+	DiskUsage   interface{} `json:"diskUsage"`
+	DiskUtil    interface{} `json:"diskUtil"`
+	DiskTotal   interface{} `json:"diskTotal"`
 	ResourceCnt interface{} `json:"resourceCnt"`
 	NodeRunning interface{} `json:"nodeRunning"`
-	NodeInfo interface{} `json:"nodeInfo"`
-	
+	NodeInfo    interface{} `json:"nodeInfo"`
 }
